Add files via upload
Here's some skill code examples that people can start with for fun or work.  Usage usually is to just type load("enc.ill")  in the

load("thefilename")       Enjoy and leave a comment if you have any questions. Look at  edagroup or edaboard in google there's a lot of good forums to help out.